--- admin.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8" />
-        <base href="/admin">
-
-        <title>Test!</title>
-    </head>
-    <body>
-        <div id="example"></div>
-
-        <!-- Dependencies -->
-        <script src="https://unpkg.com/react/umd/react.development.js"></script>
-        <script src="https://unpkg.com/react/umd/react.production.min.js"></script>
-
-        <!-- Main -->
-        <script src="/dist/dist/bundle.js"></script>
-    </body>
-</html>
