+++ index.html
@@ -4,7 +4,7 @@
         <meta charset="UTF-8" />
         <base href="/">
 
-        <title>Typescript!</title>
+        <title>TrackIT</title>
     </head>
     <body>
         <div id="root"></div>
