--- pricing.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8" />
-        <base href="/pricing">
-
-        <title>Typescript!</title>
-    </head>
-    <body>
-        <div id="example"></div>
-
-        <!-- Dependencies -->
-        <script src="./node_modules/react/umd/react.development.js"></script>
-        <script src="./node_modules/react-dom/umd/react-dom.development.js"></script>
-
-        <!-- Main -->
-        <script src="/dist/dist/bundle.js"></script>
-    </body>
-</html>
