+++ src/views/LandingPage.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { useEffect } from "react";
-import Navbar from "../Layout/NavBar";
+import Navbar from "../components/Layout/NavBar";
 import { Container, Button } from "react-bootstrap";
 import { useAuth0 } from '@auth0/auth0-react';
 
@@ -18,12 +18,7 @@ const LandingPage = () => {
 
   return (
     <div>
-      <Navbar 
-        dashboard = {false}
-        signOut={() => {
-          console.log("placeholder function");
-        }}
-      />
+      <Navbar />
       <Container fluid>
         <div className="m-5 p-5 floating-icon">
           <h1>Welcome to TrackIT</h1>
